refactor: modularize sign data and practice components
