+++ src/App.jsx
@@ -1,5 +1,7 @@
-import React, { useEffect, useMemo, useRef, useState } from 'react';
-import * as handPoseDetection from '@tensorflow-models/hand-pose-detection';
+import React, { useEffect, useMemo, useState } from 'react';
+import PracticeView from './components/PracticeView.jsx';
+import useHandsDetector from './hooks/useHandsDetector.js';
+import { ALL_SIGNS, AI_SUPPORTED } from './data/signs.js';
 
 // ============================================
 // ASL Baby Signs – Web POC (Webcam + AI + SRS)
@@ -22,59 +24,6 @@ import * as handPoseDetection from '@tensorflow-models/hand-pose-detection';
 // - Serve over HTTPS for camera permissions.
 // - Some ASL signs require motion/body/face context; this POC uses simple heuristics.
 // - You can later replace heuristics with a TF.js model to expand coverage.
-
-/*************************
- * Sign List & Metadata (plain JS)
- *************************/
-
-const ALL_SIGNS = [
-  { id: 'Help', category: 'Actions' },
-  { id: 'All Done', gloss: 'FINISH', category: 'Actions' },
-  { id: 'Please', category: 'Politeness' },
-  { id: 'More', category: 'Mealtime' },
-  { id: 'Good', category: 'Politeness' },
-  { id: 'Happy', category: 'Feelings' },
-  { id: 'Sad', category: 'Feelings' },
-  { id: 'Sleep', category: 'Routines' },
-  { id: 'Drink', category: 'Mealtime' },
-  { id: 'Eat', gloss: 'EAT/FOOD', category: 'Mealtime' },
-  { id: 'Spoon', category: 'Objects' },
-  { id: 'Bed', category: 'Objects' },
-  { id: 'Diaper', category: 'Routines' },
-  { id: 'Book', category: 'Objects' },
-  { id: 'Mommy', gloss: 'MOTHER', category: 'People' },
-  { id: 'Daddy', gloss: 'FATHER', category: 'People' },
-  { id: 'Grandma', gloss: 'GRANDMOTHER', category: 'People' },
-  { id: 'Grandpa', gloss: 'GRANDFATHER', category: 'People' },
-  { id: 'Baby', category: 'People' },
-  { id: 'Rain', category: 'Weather' },
-  { id: 'House', category: 'Places' },
-  { id: 'Car', category: 'Objects' },
-  { id: 'Stroller', category: 'Objects' },
-  { id: 'I Love You', gloss: 'ILY', category: 'Politeness' },
-  { id: 'Hug', category: 'Feelings' },
-  { id: 'Cold', category: 'Feelings' },
-  { id: 'Pain', gloss: 'HURT', category: 'Feelings' },
-  { id: 'Open', category: 'Actions' },
-  { id: 'Close', category: 'Actions' },
-  { id: 'Cry', category: 'Feelings' },
-  { id: 'Play', category: 'Actions' },
-  { id: 'Stop', category: 'Actions' },
-  { id: 'Go', category: 'Actions' },
-  { id: 'Laugh', category: 'Feelings' },
-  { id: 'Tired', category: 'Feelings' },
-  { id: 'Up', category: 'Directions' },
-  { id: 'Out', category: 'Directions' },
-  { id: 'Hold Me', category: 'Actions' },
-];
-
-// Minimal initial recognizer coverage for the POC.
-const AI_SUPPORTED = {
-  'I Love You': true,
-  More: true,
-  Help: true,
-  Stop: true,
-};
 
 /*************************
  * Spaced Repetition (SM2-lite)
@@ -348,258 +297,6 @@ function recognize(target, hands) {
  * Camera + Detector Hook (robust init)
  *************************/
 
-function useHandsDetector() {
-  const videoRef = useRef(null);
-  const canvasRef = useRef(null);
-  const [detector, setDetector] = useState(null);
-  const [ready, setReady] = useState(false);
-  const [err, setErr] = useState('');
-
-  useEffect(() => {
-    let stopped = false;
-    async function init() {
-      try {
-        if (!navigator.mediaDevices || !navigator.mediaDevices.getUserMedia) {
-          throw new Error('Camera API not available');
-        }
-        const stream = await navigator.mediaDevices.getUserMedia({
-          video: { width: 640, height: 480 },
-        });
-        if (!videoRef.current) return;
-        videoRef.current.srcObject = stream;
-        await videoRef.current.play().catch(() => {});
-
-        if (typeof navigator !== 'undefined' && !navigator.onLine) {
-          throw new Error('Network unavailable for hand detector');
-        }
-
-        const model = handPoseDetection.SupportedModels.MediaPipeHands;
-        const d = await handPoseDetection.createDetector(model, {
-          runtime: 'mediapipe',
-          modelType: 'lite',
-          solutionPath: import.meta.env.DEV
-            ? '/node_modules/@mediapipe/hands'
-            : '/hands',
-        });
-        if (stopped) return;
-        setDetector(d);
-        setReady(true);
-      } catch (e) {
-        console.error('Init error', e);
-        setErr(e && e.message ? e.message : 'Unknown error');
-        setReady(true); // default to manual practice
-        // Keep UI usable without camera/detector
-      }
-    }
-    init();
-    return () => {
-      stopped = true;
-      if (videoRef.current && videoRef.current.srcObject) {
-        const tracks = videoRef.current.srcObject.getTracks?.() || [];
-        tracks.forEach((t) => t.stop());
-      }
-      // Note: if detector exposes dispose(), consider calling it here.
-    };
-  }, []);
-
-  return { videoRef, canvasRef, detector, ready, err };
-}
-
-/*************************
- * Practice Component (optimized + manual fallback)
- *************************/
-
-function PracticeView({ target, onResult, detector, videoRef, canvasRef }) {
-  const [status, setStatus] = useState('Try the sign when you’re ready');
-  const [aiSupported, setAiSupported] = useState(!!AI_SUPPORTED[target]);
-  const stableCounter = useRef(0);
-  const rafRef = useRef(null);
-
-  // Performance: gate status updates to avoid re-render every frame
-  const lastStatus = useRef('');
-  const setStatusIfChanged = (s) => {
-    if (lastStatus.current !== s) {
-      lastStatus.current = s;
-      setStatus(s);
-    }
-  };
-
-  // Throttle detection FPS
-  const lastTsRef = useRef(0);
-  const FPS = 20; // ~20fps saves CPU/GPU
-  const FRAME = 1000 / FPS;
-
-  useEffect(() => {
-    setAiSupported(!!AI_SUPPORTED[target]);
-    stableCounter.current = 0;
-  }, [target]);
-
-  useEffect(() => {
-    let running = true;
-
-    async function loop(ts) {
-      if (!running) return;
-
-      // Pause expensive work when tab is hidden
-      if (
-        typeof document !== 'undefined' &&
-        document.visibilityState === 'hidden'
-      ) {
-        rafRef.current = requestAnimationFrame(loop);
-        return;
-      }
-
-      // FPS throttle
-      if (ts != null) {
-        const last = lastTsRef.current;
-        if (ts - last < FRAME) {
-          rafRef.current = requestAnimationFrame(loop);
-          return;
-        }
-        lastTsRef.current = ts;
-      }
-
-      if (!videoRef.current || !canvasRef.current) {
-        // Wait until video/canvas are ready (detector may be null in manual mode)
-        rafRef.current = requestAnimationFrame(loop);
-        return;
-      }
-
-      const video = videoRef.current;
-      const canvas = canvasRef.current;
-      const ctx = canvas.getContext('2d');
-      if (!ctx) {
-        rafRef.current = requestAnimationFrame(loop);
-        return;
-      }
-
-      canvas.width = video.videoWidth || 640;
-      canvas.height = video.videoHeight || 480;
-
-      ctx.save();
-      ctx.drawImage(video, 0, 0, canvas.width, canvas.height);
-
-      let hands = [];
-      if (detector) {
-        try {
-          hands = await detector.estimateHands(video, { flipHorizontal: true });
-        } catch {
-          // ignore
-        }
-      }
-
-      const canAI = aiSupported && !!detector;
-
-      // Draw landmarks if we have them
-      if (hands && hands.length) {
-        ctx.lineWidth = 2;
-        for (const hand of hands) {
-          const kps = hand.keypoints || [];
-          for (const kp of kps) {
-            if (!kp) continue;
-            ctx.beginPath();
-            ctx.arc(kp.x, kp.y, 3, 0, Math.PI * 2);
-            ctx.fillStyle = '#10b981';
-            ctx.fill();
-          }
-          const chains = [
-            [0, 1, 2, 3, 4],
-            [0, 5, 6, 7, 8],
-            [0, 9, 10, 11, 12],
-            [0, 13, 14, 15, 16],
-            [0, 17, 18, 19, 20],
-          ];
-          ctx.strokeStyle = '#16a34a';
-          for (const chain of chains) {
-            ctx.beginPath();
-            for (let i = 0; i < chain.length; i++) {
-              const p = kps[chain[i]];
-              if (!p) continue;
-              if (i === 0) ctx.moveTo(p.x, p.y);
-              else ctx.lineTo(p.x, p.y);
-            }
-            ctx.stroke();
-          }
-        }
-      }
-
-      let recognized = null;
-      if (canAI) {
-        recognized = recognize(target, hands);
-      }
-
-      if (
-        canAI &&
-        recognized &&
-        recognized.label === target &&
-        recognized.confidence > 0.8
-      ) {
-        stableCounter.current += 1;
-        setStatusIfChanged(
-          `Detected: ${recognized.label} (${recognized.confidence.toFixed(2)})`
-        );
-      } else if (canAI) {
-        stableCounter.current = Math.max(0, stableCounter.current - 1);
-        setStatusIfChanged('Listening… try the sign');
-      } else {
-        // Manual fallback
-        setStatusIfChanged('Manual practice: use buttons below.');
-      }
-
-      if (canAI && stableCounter.current >= 8) {
-        running = false;
-        setStatusIfChanged('Great job! ✔ Recognized');
-        setTimeout(() => onResult(true), 350);
-        ctx.restore();
-        return;
-      }
-
-      ctx.restore();
-      rafRef.current = requestAnimationFrame(loop);
-    }
-
-    rafRef.current = requestAnimationFrame(loop);
-    return () => {
-      running = false;
-      if (rafRef.current) cancelAnimationFrame(rafRef.current);
-    };
-  }, [detector, target, aiSupported, onResult, videoRef, canvasRef]);
-
-  const canAI = aiSupported && !!detector;
-
-  return (
-    <div className="w-full space-y-3">
-      <div className="text-sm text-gray-500">
-        {canAI ? 'AI check enabled' : 'Manual check (camera/AI unavailable)'}
-      </div>
-      <div className="relative w-full max-w-2xl aspect-video rounded-2xl overflow-hidden shadow">
-        <video ref={videoRef} className="hidden" playsInline muted />
-        <canvas ref={canvasRef} className="w-full h-full bg-black" />
-      </div>
-      <div className="text-base font-medium">{status}</div>
-      {!canAI && (
-        <div className="flex gap-2">
-          <button
-            onClick={() => onResult(true)}
-            className="px-4 py-2 rounded-xl bg-emerald-600 text-white hover:bg-emerald-700"
-          >
-            Mark Correct
-          </button>
-          <button
-            onClick={() => onResult(false)}
-            className="px-4 py-2 rounded-xl bg-rose-600 text-white hover:bg-rose-700"
-          >
-            Mark Again
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-/*************************
- * Dev Tests (SRS + Safety)
- *************************/
 
 
 /*************************
@@ -880,6 +577,7 @@ export default function App() {
                   detector={detectorReady ? detector : null}
                   videoRef={videoRef}
                   canvasRef={canvasRef}
+                  recognize={recognize}
                 />
 
                 <div className="space-y-3">
