+++ src/data/signs.js
@@ -1,0 +1,48 @@
+export const ALL_SIGNS = [
+  { id: 'Help', category: 'Actions' },
+  { id: 'All Done', gloss: 'FINISH', category: 'Actions' },
+  { id: 'Please', category: 'Politeness' },
+  { id: 'More', category: 'Mealtime' },
+  { id: 'Good', category: 'Politeness' },
+  { id: 'Happy', category: 'Feelings' },
+  { id: 'Sad', category: 'Feelings' },
+  { id: 'Sleep', category: 'Routines' },
+  { id: 'Drink', category: 'Mealtime' },
+  { id: 'Eat', gloss: 'EAT/FOOD', category: 'Mealtime' },
+  { id: 'Spoon', category: 'Objects' },
+  { id: 'Bed', category: 'Objects' },
+  { id: 'Diaper', category: 'Routines' },
+  { id: 'Book', category: 'Objects' },
+  { id: 'Mommy', gloss: 'MOTHER', category: 'People' },
+  { id: 'Daddy', gloss: 'FATHER', category: 'People' },
+  { id: 'Grandma', gloss: 'GRANDMOTHER', category: 'People' },
+  { id: 'Grandpa', gloss: 'GRANDFATHER', category: 'People' },
+  { id: 'Baby', category: 'People' },
+  { id: 'Rain', category: 'Weather' },
+  { id: 'House', category: 'Places' },
+  { id: 'Car', category: 'Objects' },
+  { id: 'Stroller', category: 'Objects' },
+  { id: 'I Love You', gloss: 'ILY', category: 'Politeness' },
+  { id: 'Hug', category: 'Feelings' },
+  { id: 'Cold', category: 'Feelings' },
+  { id: 'Pain', gloss: 'HURT', category: 'Feelings' },
+  { id: 'Open', category: 'Actions' },
+  { id: 'Close', category: 'Actions' },
+  { id: 'Cry', category: 'Feelings' },
+  { id: 'Play', category: 'Actions' },
+  { id: 'Stop', category: 'Actions' },
+  { id: 'Go', category: 'Actions' },
+  { id: 'Laugh', category: 'Feelings' },
+  { id: 'Tired', category: 'Feelings' },
+  { id: 'Up', category: 'Directions' },
+  { id: 'Out', category: 'Directions' },
+  { id: 'Hold Me', category: 'Actions' },
+];
+
+export const AI_SUPPORTED = {
+  'I Love You': true,
+  More: true,
+  Help: true,
+  Stop: true,
+};
+
